requirements.txt
numpy<2.0.0
streamlit
sqlalchemy
psycopg2-binary
pandas
openpyxl
plotly
python-dotenv
TA-Lib